data append to crypto table" "

+++ src/db/db_init.sql
@@ -8,13 +8,13 @@ END $$;
 \c crypto;
 
 CREATE TABLE IF NOT EXISTS Crypto (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name VARCHAR(20) NOT NULL,
     ticker VARCHAR(5) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS Time (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     crypto_id INT NOT NULL REFERENCES Crypto(id),
     timestamp TIMESTAMP,
     year INT,
@@ -26,9 +26,9 @@ CREATE TABLE IF NOT EXISTS Time (
 );
 
 CREATE TABLE IF NOT EXISTS Price (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     crypto_id INT NOT NULL REFERENCES Crypto(id),
-    time_id INT NOT NULL REFERENCES Time(id),
+    time_id BIGINT NOT NULL REFERENCES Time(id),
     currency VARCHAR(20),
     open DECIMAL,
     close DECIMAL,
